Separate Utils - preparation for SDK, new splash
