+++ studio/src/components/Editor/Context/EditorContext.tsx
@@ -1,101 +1,40 @@
 import React from 'react';
 
-import { EditorModel } from '@utils/models/models/EditorModel';
-import { changeSelection } from '@utils/seleciton/selection';
-
-import * as GL from '@bananagl/bananagl';
+import { ViewContext, ViewContextProps } from '@utils/components/Context';
 
 export enum EditingStage {
     Transform,
     Table,
 }
 
-interface EditorContextProps {
+interface EditorContextProps extends ViewContextProps {
     processing: boolean;
     setProcessing: React.Dispatch<React.SetStateAction<boolean>>;
-    scene: GL.Scene;
-    renderer: GL.Renderer;
-    activeView: number;
-    models: EditorModel[];
-    setModels: React.Dispatch<React.SetStateAction<EditorModel[]>>;
     loadingStatus: string;
     setLoadingStatus: React.Dispatch<React.SetStateAction<string>>;
-    selectedModel: EditorModel | null;
-    selectedSubmodels: number[];
-    select: (
-        model: EditorModel | null,
-        submodelIDs?: number[],
-        toggle?: boolean,
-        extend?: boolean
-    ) => void;
     editingStage: EditingStage;
     setEditingStage: React.Dispatch<React.SetStateAction<EditingStage>>;
 }
 
 export const EditorContext = React.createContext<EditorContextProps>({} as EditorContextProps);
 
-export function ContextComponent(props: { children: React.ReactNode }) {
-    const [renderer] = React.useState(new GL.Renderer());
-    const [scene] = React.useState(new GL.Scene());
-    const [models, setModels] = React.useState<EditorModel[]>([]);
-    const [selectedModel, setSelectedModel] = React.useState<EditorModel | null>(null);
-    const [selectedSubmodels, setSelectedSubmodels] = React.useState<number[]>([]);
-    const [editingStage, setEditingStage] = React.useState<EditingStage>(EditingStage.Transform);
+export function EditorContextComponent(props: { children: React.ReactNode }) {
+    const ctx = React.useContext(ViewContext);
 
+    const [editingStage, setEditingStage] = React.useState<EditingStage>(EditingStage.Transform);
     const [loadingStatus, setLoadingStatus] = React.useState<string>('');
     const [processing, setProcessing] = React.useState(false);
-
-    React.useEffect(() => {
-        const onChange = () => {
-            const copy = scene.objects.filter((obj) => obj instanceof EditorModel) as EditorModel[];
-            setModels(copy);
-            if (selectedModel !== null && !copy.includes(selectedModel)) setSelectedModel(null);
-        };
-
-        scene.addChangeListener(onChange);
-
-        return () => {
-            scene.removeChangeListener(onChange);
-        };
-    }, [scene, selectedModel]);
-
-    //selection callback, still a bit hacky but nicer than before
-    function select(
-        model: EditorModel | null,
-        submodelIDs: number[] = [],
-        toggle: boolean = false,
-        extend: boolean = false
-    ) {
-        submodelIDs = changeSelection(
-            selectedModel,
-            model,
-            selectedSubmodels,
-            submodelIDs,
-            toggle,
-            extend
-        );
-
-        setSelectedModel(model);
-        setSelectedSubmodels(submodelIDs);
-    }
 
     return (
         <EditorContext.Provider
             value={{
                 processing,
                 setProcessing,
-                scene,
-                renderer,
-                activeView: 0,
-                models,
-                setModels,
                 loadingStatus,
                 setLoadingStatus,
-                selectedModel,
-                selectedSubmodels,
-                select,
                 editingStage,
                 setEditingStage,
+                ...ctx,
             }}
         >
             {props.children}
