--- studio/src/components/Editor/Context/TransformContext.tsx
@@ -1,46 +1,0 @@
-import { vec3 } from 'gl-matrix';
-import React from 'react';
-
-interface TransformContextProps {
-    camTargetZ: number;
-    setCamTargetZ: React.Dispatch<React.SetStateAction<number>>;
-    minShade: number;
-    setMinShade: React.Dispatch<React.SetStateAction<number>>;
-    maxShade: number;
-    setMaxShade: React.Dispatch<React.SetStateAction<number>>;
-    gridVisible: boolean;
-    setGridVisible: React.Dispatch<React.SetStateAction<boolean>>;
-    globalShift: vec3 | null;
-    setGlobalShift: React.Dispatch<React.SetStateAction<vec3 | null>>;
-}
-
-export const TransformContext = React.createContext<TransformContextProps>(
-    {} as TransformContextProps
-);
-
-export function TransformContextComponent(props: { children: React.ReactNode }) {
-    const [camTargetZ, setCamTargetZ] = React.useState<number>(0);
-    const [minShade, setMinShade] = React.useState<number>(0);
-    const [maxShade, setMaxShade] = React.useState<number>(10);
-    const [gridVisible, setGridVisible] = React.useState<boolean>(true);
-    const [globalShift, setGlobalShift] = React.useState<vec3 | null>(null);
-
-    return (
-        <TransformContext.Provider
-            value={{
-                camTargetZ,
-                setCamTargetZ,
-                minShade,
-                setMinShade,
-                maxShade,
-                setMaxShade,
-                gridVisible,
-                setGridVisible,
-                globalShift,
-                setGlobalShift,
-            }}
-        >
-            {props.children}
-        </TransformContext.Provider>
-    );
-}
