+++ studio/src/components/Editor/main.tsx
@@ -2,9 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 
-import { ContextComponent } from './Context/EditorContext';
+import { ViewContextComponent } from '@utils/components/Context';
+
+import { EditorContextComponent } from './Context/EditorContext';
 import { TablesContextComponent } from './Context/TableContext';
-import { TransformContextComponent } from './Context/TransformContext';
 import { ModelEditor } from './Editor';
 import { ErrorPage } from './Utils/Error';
 import './index.css';
@@ -19,12 +20,12 @@ const router = createBrowserRouter([
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
     <React.StrictMode>
-        <ContextComponent>
-            <TransformContextComponent>
+        <ViewContextComponent>
+            <EditorContextComponent>
                 <TablesContextComponent>
                     <RouterProvider router={router} />
                 </TablesContextComponent>
-            </TransformContextComponent>
-        </ContextComponent>
+            </EditorContextComponent>
+        </ViewContextComponent>
     </React.StrictMode>
 );
