+++ studio/src/components/Elements/MenuButton.tsx
@@ -1,12 +1,9 @@
 import clsx from 'clsx';
 import React from 'react';
 
-import { Tooltip } from './Tooltip';
-
 interface MenuButtonProps {
     children: React.ReactNode;
     tipTitle?: string;
-    tipPosition?: 'top' | 'left' | 'right' | 'bottom';
     active?: boolean;
     onClick?: () => void;
 }
@@ -15,20 +12,15 @@ export function MenuButton(props: MenuButtonProps) {
     return (
         <button
             className={clsx(
-                'text-xs last:rounded-r-md first:rounded-l-md transition-colors outline-none focus:outline-none border-y first:border-l last:border-r',
+                'text-xs last:rounded-r-md first:rounded-l-md transition-colors outline-none focus:outline-none border-y first:border-l last:border-r p-2',
                 props.active
                     ? 'text-amber-600 bg-amber-100 hover:bg-amber-200'
                     : 'text-neutral-600 bg-white hover:bg-neutral-300'
             )}
             onClick={props.onClick}
+            title={props.tipTitle}
         >
-            {props.tipTitle && props.tipPosition ? (
-                <Tooltip content={props.tipTitle} position={props.tipPosition}>
-                    <div className="p-2">{props.children}</div>
-                </Tooltip>
-            ) : (
-                <div className="p-2">{props.children}</div>
-            )}
+            {props.children}
         </button>
     );
 }
