+++ studio/src/utils/models/EditorModelShader.ts
