+++ studio/src/utils/modifiers/computeNormals.ts
